+++ ReleaseNotes.md
@@ -11,6 +11,8 @@
 - Blank books
   - new section in library tab with a list of physical books
   - Default state of books created as enchantments
+- [Warning] Compatibility with mono-topic books from V9 has been removed, migrate any old data before updating.
+- It is now possible to write a new topic to an existing book.
 
 ### Bug fixes
 
@@ -21,6 +23,8 @@
   - dependants are again displayed in the Inhabitants section
   - Points are readonly for now and based on the role.
 - Confirmation of book topic deletion is taken into account again (ie: saying no will have an effect)
+- Fixed the book topic filter
+- Ability keys categories are again more distinguishable
 
 ## 2.2.4.4, Pralix, filia Guorna the Fetid